+++ kernel-spark/src/test/java/io/delta/kernel/spark/read/SparkScanTest.java
@@ -3,7 +3,7 @@ package io.delta.kernel.spark.read;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import io.delta.kernel.spark.SparkDsv2TestBase;
-import io.delta.kernel.spark.table.SparkTable;
+import io.delta.kernel.spark.catalog.SparkTable;
 import java.io.File;
 import java.lang.reflect.Field;
 import java.util.ArrayList;
